requirements.txt
google-images-download==2.8.0
Pillow==6.1.0
selenium==3.141.0
urllib3==1.25.3
